+++ src/base/clptur.f90
@@ -157,7 +157,7 @@ double precision hbord(nfabor),theipb(nfabor)
 ! Local variables
 
 integer          ifac, iel, isou, ii, jj, kk
-integer          iscal
+integer          iscal, clsyme
 integer          modntl
 integer          iuntur, f_dim
 integer          nlogla, nsubla, iuiptn
@@ -170,7 +170,7 @@ double precision uiptn, uiptmn, uiptmx
 double precision uetmax, uetmin, ukmax, ukmin, yplumx, yplumn
 double precision tetmax, tetmin, tplumx, tplumn
 double precision uk, uet, nusury, yplus, dplus
-double precision sqrcmu, clsyme, ek
+double precision sqrcmu, ek
 double precision xnuii, xnuit, xmutlm
 double precision rcprod
 double precision hflui, hint, pimp, qimp
@@ -681,37 +681,10 @@ do ifac = 1, nfabor
       eloglo(3,2) = -rnz
       eloglo(3,3) =  t2z
 
-      ! --> Commpute alpha(6,6)
+      ! Compute Reynolds stress transformation matrix
 
-      ! Let f be the center of the boundary faces and
-      !   I the center of the matching cell
-
-      ! We noteE Rg (resp. Rl) indexed by f or by I
-      !   the Reynolds Stress tensor in the global basis (resp. local)
-
-      ! The alpha matrix applied to the global vector in I'
-      !   (Rg11,I'|Rg22,I'|Rg33,I'|Rg12,I'|Rg13,I'|Rg23,I')t
-      !    must provide the values to prescribe to the face
-      !   (Rg11,f |Rg22,f |Rg33,f |Rg12,f |Rg13,f |Rg23,f )t
-      !    except for the Dirichlet boundary conditions (added later)
-
-      ! We define it by computing Rg,f as a function of Rg,I' as follows
-
-      !   RG,f = ELOGLO.RL,f.ELOGLOt (matrix products)
-
-      !                     | RL,I'(1,1)     B*U*.Uk     C*RL,I'(1,3) |
-      !      with    RL,f = | B*U*.Uk       RL,I'(2,2)       0        |
-      !                     | C*RL,I'(1,3)     0         RL,I'(3,3)   |
-
-      !             with    RL,I = ELOGLOt.RG,I'.ELOGLO
-      !                     B = 0
-      !              and    C = 0 at the wall (1 with symmetry)
-
-      ! We compute in fact  ELOGLO.projector.ELOGLOt
-
-      clsyme=0.d0
-      call clca66 (clsyme , eloglo , alpha)
-      !==========
+      clsyme = 0
+      call turbulence_bc_rij_transform(clsyme, eloglo, alpha)
 
     endif